+++ R/bind_probs.R
@@ -1,4 +1,4 @@
-#' Stack Flexible-Probabilities
+#' Stack Flexible Probabilities
 #'
 #' This function mimics `dplyr` \code{\link[dplyr]{bind}}. It's useful if you
 #' have different `ffp` objects and want to stack them in the `tidy` (long) format.
@@ -13,7 +13,7 @@
 #'   \item `key` (a \code{factor}) that keeps track of the `ffp` inputs as separated objects.
 #' }
 #'
-#' @seealso \code{\link{crisp}} \code{\link{smoothing}} \code{\link{kernel_normal}}
+#' @seealso \code{\link{crisp}} \code{\link{exp_smoothing}} \code{\link{kernel_normal}}
 #' \code{\link{kernel_entropy}} \code{\link{double_decay}}
 #'
 #' @export
@@ -22,8 +22,8 @@
 #' library(ggplot2)
 #' library(dplyr, warn.conflicts = FALSE)
 #'
-#' x <- smoothing(EuStockMarkets, 0.001)
-#' y <- smoothing(EuStockMarkets, 0.002)
+#' x <- exp_smoothing(EuStockMarkets, 0.001)
+#' y <- exp_smoothing(EuStockMarkets, 0.002)
 #'
 #' bind_probs(x, y)
 #'