+++ NEWS.md
@@ -2,7 +2,8 @@
 
 * The package now is built upon S3 `vctrs`, instead of `tibbles`;
 * Functions `bootstrap_scenarios()`, `scenario_density()` and `bind_probs()` only accepts objects of `ffp` class;
-* Added `ffp` method to `ggplot2::autoplot()`
+* Added `ffp` method to `ggplot2::autoplot()`;
+* Exponential Smoothing is now computed with `exp_smoothing()`. 
 
 # ffp 0.1.9000